Add React's Synthetic event types for flow to list of globals
Summary:
We'd like to use flow to do static type checking.  Flow requires that
we type the event params to React event handlers handlers as SyntheticEvent
or one of its subtypes.  Unfortunately, there's no way to require these
types so eslint flags them as being undefined which they would be if
babel didn't strip them out for us.

This diff adds SyntheticEvent and all of its subtypes to the list of
globals in eslintrc.shared.

https://phabricator.khanacademy.org/D32066 adds the same list of event
types to `_NON_REQUIREABLE_GLOBALS` in webapp.

Test Plan:
- remove the @Nolint from https://phabricator.khanacademy.org/D32001
- run ka-lint src/widgets/interaction/element-container.jsx
- see that there are no lint violations

Reviewers: csilvers, jared

Reviewed By: csilvers, jared

Differential Revision: https://phabricator.khanacademy.org/D32033