Add a max-len linter for JavaScript files
Summary:
We've been discussing this for a long time, and I _think_ we're pretty
agreed on 1k being a safe file-length limit.

I'd be happy to bump this down into the neighborhood of 400, but not
before sending out a survey. I figure this diff can be a launch-point
for discussion.

Test Plan:
(from webapp/)

- `make lint`

Tests should fail with:
`Emax-lines File must be at most 1000 lines long. It's x lines long.`

Before shipping this, I'll update our `eslint-disable` lines in
webapp, perseus, etc.

Reviewers: emily, mattdunnrankin, amyciavolino, john, kevinb

Reviewed By: john, kevinb

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D32565