Fix runlint.py to fully support **, and to not need trailing / anymore.
Summary:
Both of these constraints are non-obvious, and the latter especially
bit us from time to time (people would add directories to a blacklist
and not realize they needed the trailing /).

Supporting `**` is just a matter of code -- copied from kake, so it's
actually better-tested than the test-suite would lead you to believe.
Supporting dirs makes blacklist-checking a bit slower than it used to
be, but for all but the largest changelists it shouldn't matter that
much.

Test Plan: ./runlint_test.py

Reviewers: emily, mroth

Reviewed By: mroth

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D32987