Add a chdir so the linter runs properly in other directories.
runtests.py (in webapp) tries to run the linter in webapp-root, but
the linter is only happen when it's run in a subdirectory of where the
blacklist file lives.  The easiest way to fix that is to just chdir.
I use addCleanup so hopefully it's safe.

Auditors: mroth

Test plan:
Ran
   cd ~ && ~/khan/devtools/khan-linter/runlint_test.py
and the tests passed.