Add eslint-plugin-graphql as a potential plugin
Summary:
This patch is just the result of running `npm install
eslint-plugin-graphql`. We'd like to use this eslint plugin in webapp,
but I've had trouble configuring eslint to use a npm package installed
in webapp. This should not require any other users of khan-lint to run
graphql testing, but will give them the option to do so if they choose.

Test Plan:
Will run `ka-lint` on a couple filetypes in webapp to ensure I didn't
break anything.

Reviewers: john, csilvers

Reviewed By: csilvers

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D34141