Update khan-linter so it works for both python2 and python3
Summary:
Note: this diff is split in two parts.  This diff contains the changes to non-vendored code so that they're easier to review.  See D35060 for the changes to vendored code.

`khanalytics` is written in python3 and is soon going to be used in production.
Up until now, we've just been using in-editor linting, but we'd like to move to
using khan-linter to keep style consistent with our other repositories and to
make running linters as part of committing and arc easier.

Unfortunately, khan-linter is not compatible with python3 in two ways:
1. It can't lint python3.4+ because pyflakes and pep8 are pinned on *very*
   old versions.
2. It can't be run under python3 because the linter scripts themselves are
   written in python2.

This commit fixes both of these issues.  For (1), I upgraded pep8 and flake8 to
the latest versions and vendored them in.  This required creating separate
vendor/py2 and vendor/py3 directories.  I also configured the makefile to wipe
out the vendor directory when doing `make vendor_deps` because otherwise you
can get residual old versions in there depending on what you had there before,
and the whole point of the vendor dir is to provide a clean consistent
environment for everyone.  Doing all this of course introduced a bunch of lint
problems in all our other repos because there's new linter rules.  I was
originally planning to disable the new rules, but there's not that many new
problems, and some of them are really nice to have (like variables that escape
list comprehensions redefining existing ones and duplicate dict keys), so
there's a parallel commit in webapp that just fixes all of the lint issues.

For (2), I vendored in `six` (in the base of the repo, not the vendor dir,
since we use `six` to determine which vendor dir to import) and used it to
implement the few things that needed to change (mostly `print` and the
iteration functions).  I also upgraded the old `except Exception, name` syntax
to `except Exception as name` everywhere, since that's recommended even under
python2.

Note that I haven't upgraded the git hook for the linter yet.  I'll do that
separately.

I also added a setup.cfg file, which is a workaround for a bug in distutils
that precludes using a homebrew-installed python on osx.  Normally we work
around this by using the system python, but there is no system python3.

Test Plan:
- in a python3.6 virtualenv in the khanalytics repo, run:
```
python ~/khan/devtools/khan-linter/runlint.py core gcloud integrations kubernetes
```
- verify the linter completes successfully and shows a whole bunch of lint
  errors I need to fix.
- in a python2.7 virtualenv in the webapp repo, run:
```
`make lint`
```
- verify the linter completes successfully and finds no errors (after the
  corresponding webapp fixing the new ones)

Reviewers: csilvers, mroth

Reviewed By: csilvers

Subscribers: kevinb, ragini, tom

Differential Revision: https://phabricator.khanacademy.org/D35059