Fix the commit hooks to work on python2 and python3
Summary:
In D35059 and friends, I upgraded us to that we can run khan-linter on (and
using) either python2 or python3.  However, our commit hooks would still fail
when run under python3.  This commit modifies the hooks so that they can run
under either version.

There was a bit of tricky unicode vs. bytes going on here with the output of
the git commands.  From poking around the git documentation a bit it looks like
we should expect to get filenames utf-8 encoded.  So I think the right thing to
do is to strip off the delimiters and then decode with utf-8.

Test Plan:
- in the khanalytics repo, in a python3 virtualenv, try to make a commit that
  introduces a lint error, see that our hook prevents the commit and reports
  the error (previously it was impossible to run `git commit` from a python3
  virtualenv if hooks were enabled)
- in the webapp repo, in a python2 virualenv, try to make a commit that
  introduces a lint error, see that our hook prevents the commit and reports
  the error

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: tom, kevinb

Differential Revision: https://phabricator.khanacademy.org/D35086