Add a test that imports runlint under both python2 and python3
Summary:
khan-linter now runs under both python2 and python3, but it's going to be easy
to break that.  This commit adds a test that checks that runlint.py can be
imported in both python2 and python3.  This won't catch everything, but will
prevent a few common issues, like print statements without parens and octal
literals for permissions.

Test Plan:
- `python -m unittest python_version_compat_test`
- verify it passes
- add a `print 'broken'` to runlint.py
- repeat, and verify the python3 test fails

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: tom, kevinb

Differential Revision: https://phabricator.khanacademy.org/D35087