Move add_lint_ignore_lines.py from webapp to the linter
Summary:

In D35072, I added a script to the webapp repo to add python lint ignore lines
for the purpose of upgrading the python linter.  @kevinb is adding a similar
javascript veresion to the khan-linter repo, so this commit moves the python
version from the webapp repo to the linter repo.

In addition to the moving, I had to modify the import of khan linter, since
it's no longer the version from third_party but from the current directory.

Test Plan:
- `add_lint_ignore_lines.py .`
- verify I see a message for each of the first-party files in this directory
  saying there's no lint errors

Auditors: kevinb, csilvers