Update pyflakes to support MyPy annotations.
Summary:
In order to support the nicer Python3-style type annotations, we need a more
up-to-date version of pyflakes.

We also need to ignore the "multiple statements on one line" lint error E701
which it seems erroneously fires on simple typed variable declarations, but
leaving that for a separate diff as it's less obvious it's something we want to
keep longer-term.

Test Plan:
In the `khanalytics` repo, run `make lint`
Also tested `make lint` in `webapp`

Reviewers: colin, ragini, csilvers

Reviewed By: colin, ragini, csilvers

Subscribers: #infrastructure, csilvers, ragini, kevinb

Differential Revision: https://phabricator.khanacademy.org/D35522