Set working directory in python_version_compat_test.py
Summary:
These tests were passing when run from within `khan-linter`, but failing when run via `webapp`'s
`tools/runtests.py`. The issue seems to be that when run from `webapp`, the current working
directory in the `subprocess` calls was `~/khan/webapp`, and so `runlint.py` couldn't be found.
Setting the working directory explicitly to match the test path resolves the issue.

Test Plan:
- From `khan-linter`, run: `python -m unittest python_version_compat_test`.
    - Verify that both tests pass without error.
- From: `~/khan/webapp`, run: `./tools/runtests.py third_party/khan-linter-src/python_version_compat_test.py`.
    - Verify that both tests pass without error.
- Introduce some non-Python 3 compatible into `runlint.py` (e.g., `print 'broken'`).
- Re-run the tests from the two locations listed above.
    - Verify that the Python 3 test fails in both cases, but the Python 2 test passes.

Reviewers: colin

Reviewed By: colin

Subscribers: csilvers, kevinb

Differential Revision: https://phabricator.khanacademy.org/D35546