Ignore E701 (multiple statements on one line) for python3
Summary:
In D35522 @tom upgraded the linter again.  In order to get pep8 to deal with
python3-style type annotations, we had to ignore one of the rules, E701, but
it'd be nice to keep this on where possible.

This commit ignores E701 only in python3.  The correct fix is to use
pycodestyle instead of pep8, since it deals with these out of the box, so I've
added a TODO to do that.

Test Plan:
- in the khanalytics repository, `make lint`, verify there are no lint errors
  reported

Reviewers: tom, ragini, csilvers

Reviewed By: ragini

Subscribers: tom, kevinb

Differential Revision: https://phabricator.khanacademy.org/D35557