Report eslint errors better
Summary:
Eslint is printing errors to stdout, which results in the cryptic error "list
index out of range" when running `arc lint` & friends.
This will make it so errors are actually bubbled up.

Test Plan:
with the graphql schema broken, run `arc lint`, and see that you actually get
a useful error.

Reviewers: kevinb, benkraft, csilvers

Reviewed By: csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D35887