Have _maybe_pull() fail gracefully if we can't read from the git remote.
Summary:
If github is down, or the user is trying to run lint on a place
without an internet connection, the linter will fail (due to `git
pull` failing) unless they know to run with `--no-auto-pull`).

Now we try to connect to the git remote before running our git
commands, and gracefully handle the case that that connection fails.
It means every ka-lint invocation takes 3 seconds to do nothing, but
that's better than it was before.  (If the git-remote access fails
right away, then even better.)

Fixes https://app.asana.com/0/31965416896056/358396296537069/f

Test Plan:
I changed the argument to _run_command_with_timeout to be
   ['sleep', '10']
and saw
   % ka-lint Makefile
   [...pause...]
   Non-fatal error: not updating the khan-linter repo: Timeout running (['sleep', '10'],)/{'stderr': <open file '/dev/null', mode 'wb' at 0x2b5001b62a50>, 'stdout': <open file '/dev/null', mode 'wb' at 0x2b5001b62a50>}

I then changed it to be
   ['git', 'ls-remotee', 'origin', 'master']
and saw
   % ka-lint Makefile
   Non-fatal error: not updating the khan-linter repo: git ls-remote returned 1

Reviewers: colin

Subscribers: kevinb, jared

Differential Revision: https://phabricator.khanacademy.org/D36338