Add prettier support to khan-linter
Summary:
This diff adds prettier support to khan-linter. This comes in a few forms:

1) The eslint style rules have been updated to align with prettier's rules. They tend to align pretty well with how we have things right now in webapp, which is good. I also removed the `max-len` rule as we'll be using `prettier` to enforce line length limits. Having an additional line-length rules is confusing and only serves to create confusion. In practice, in our code base, we're ok with long lines (like strings - such as URLs, SVG paths, etc.) but not long code. Prettier fixes the issue of long code and leaves the other pieces intact without creating frustration (we'll no longer have to add `// @Nolint` to all the long lines in our files).
2) The prettier plugin was added to eslint, along with a new `eslintrc.prettier` config which means that we can use eslint to automatically fix prettier syntax changes with eslint. For example we can `eslint --fix some-js-file.js` and eslint will automatically update the file with the new syntax, outputting any additional lint errors. Additionally, if not run with `--fix` eslint will show errors for all of the differences that prettier would fix.
3) Support for `arc diff`/`arc lint` auto-patching the prettier fixes was added. If you run `arc diff` against a piece of (un-pretty) JavaScript code a patch will be provided which you can opt to apply to your file (or diff, as the case may be).

I should note that I also had to update `eslint` and `babel-eslint` to support the prettier integration.

Test Plan:
I ran `ka-lint javascript` in webapp and confirmed that it passed.

I then changed webapp's `.eslintrc` file to point to `eslintrc.prettier` and ran `ka-lint` against a JavaScript file and saw that it output a bunch of syntax errors and diff which I could apply to the file to update it to match the prettier style. Accepting it updated the file and the linter passed.

I also made a change to a JavaScript file and ran `arc lint`. It output a list of prettier syntax errors and then gave me the option to apply a patch to my file, updating it. Doing so updated it to match the prettier style.

Reviewers: csilvers, kevinb, emily

Subscribers: