Disable `--extra-linter` when running runlint.py.
pep8, which this script is concerned with, is a standard linter.  The
extra linters are not going to add any new pep8 errors!  By turning
them off this script runs an order of magnitude faster.

Auditors: colin

Test plan:
Ran
   ~/khan/devtools/khan-linter/add_pep8_disable_lines.py
in webapp-root and had it run both correctly and quickly.