Fix some issues with prettier in khan-linter
Summary: This diff removes the hint for prettier (it was causing the auto-patch mechanism to break). Additionally, I found that the 80 character line limit stuff was causing Prettier to get confused, and so I removed it (which makes sense since Prettier is in control of that logic in teh first place).

Test Plan:
I ran

I then added a .eslintrc file to javascript/class-package:

```
{
    "extends": [
        "../../third_party/khan-linter-src/eslintrc.prettier",
        "../../eslintrc.graphql.js",
    ]
}
```

and ran the following:

```
arc lint javascript/class-package/class.jsx
```

It proposed a large patch, I accepted it, and the linter passed.

Reviewers: csilvers, john

Reviewed By: csilvers

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D40834