Use tools/runlint.sh as the default extra-linter, not runlint.py.
Summary:
We run the extra-linter as a subprocess, so there's no need for it to
be python.  Presumably it's easier to drive an extra linter from
shell-land than python-land.

In fact, we've recently changed webapp to use a runlint.sh wrapper
around its own runlint.py.  So this change will start using that
wrapper automatically.

I checked all the repos I could think of, including mobile, perseus,
even some opensource repos, and none of them define a tools/runlint.py
so this should be safe to change.  (The only repo that did is
khan-exercises, but we don't use that anymore.)

Test Plan: Fingers crossed

Reviewers: benkraft, riley

Reviewed By: benkraft

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D40866