Add a pre-push lint hook to khan-linter (not yet installed in webapp)
Summary:
As we integrate Prettier auto-JS-formatting into webapp (and into our linter), we need to change our general linting workflow. Currently, we use a git hook to reject commits whose changes fail linting - but, once Prettier is installed, changes to JS are generally //expected// to fail linting.

So, to improve the linting UX, we plan to stop running the linter at commit time. Instead, we'll continue to lint at `arc diff` time (which has a nice autofixer UI, so Prettier changes will be easy to accept!), and we'll add a new pre-push hook to continue to make it difficult to accidentally push unlinted changes via other workflows (e.g. post-push review).

In this change, we implement the pre-push hook logic in `khan-linter`. It looks for commits that will be pushed to a remote repo for the first time, and lints the files added/modified in each such commit. It is invoked via `~/khan/devtools/khan-linter/githook.py --hook=pre-push remote_name remote_location`.

This change does not affect webapp, or the ka-clone tools that install git hooks in webapp. We also take care not to break existing webapp git hooks, by leaving the `commit-msg` hook invocation syntax unchanged.

In a subsequent change, we'll update `ka-clone` to install a corresponding pre-push git hook in webapp, which will invoke this `khan-linter` script. After that, we'll update the `commit-msg` handler in `khan-linter` to no longer lint files, and only lint the commit message.

Test Plan:
Add a temporary line to the pre-push handler in `githook.py`, to print the set of linted files. Then, run `test_githook_pre_push.sh`, and manually confirm that the actual set of linted files always matches the expected set.

Additionally, I installed this as a pre-push git hook in `webapp` as well, and did some very basic testing to confirm that integration works as expected.

Reviewers: csilvers, john, kevinb

Reviewed By: csilvers, john

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D41012