[khan-linter] fix error messages that contain Unicode
Whoops, this `khan-linter` line will crash if you try to interpolate a Unicode error string into the ascii format string. In this change, we update the format string to be a Unicode string as well.

Test Plan:
Run the lint command that was previously crashing, and see that it no longer crashes.

Auditors: john