Skip symlinks when linting.
If `a` is a symlink to `b`, we should just lint one of the two files, not
both.  We pick `b` which is the canonical location.  Plus, that way if
you want to blacklist it or something you only have to list `b`, not
also list `a`.

Auditors: benkraft

Test plan:
   % ka-lint -v tools/runlint.py
   Considering /home/csilvers/khan/webapp/tools/runlint.py: blacklist None... skipping (is a symlink)

   % ka-lint -v testing/runlint.py
   Considering /home/csilvers/khan/webapp/testing/runlint.py: blacklist None... LINTING