Only remove eslint-disable lines if they're in the header
Summary:
Our tool to update top-of-file `eslint-disable` comments first strips the existing `eslint-disable` header comments, then runs ESLint, then re-adds comments for any uncaught error classes.

However, this includes `eslint-disable` comments that weren't generated by the tool - for example, suppressing a particular code block (`/* eslint-disable X */ ...lines... /* eslint-enable X */`), or the new suppression in `ka-define.js` intentionally targeting Prettier.

In this change, we update the comment-stripping code to //only// remove `eslint-disable` comments that seem to be part of the top-of-file header, and therefore seem to be generated by this tool. Once we find a line that is not part of the top-of-file header, all subsequent lines are accepted.

Test Plan:
```
./update_eslint_disable_lines.js ~/khan/webapp/javascript/corelibs-package/ka-define.js
```

Confirm that the top-of-file header is updated (we stop suppressing a redundant formatting rule), but the `/* eslint-disable prettier/prettier */` line stays intact.

Reviewers: kevinb, john, csilvers

Reviewed By: john, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D41196