Accept more arguments in update_eslint_disable_lines.js
Summary:
This is our khan-linter utility script to update the top-of-line `eslint-disable` comments in JS files. It currently accepts a single directory as its argument.

However, we'd like to integrate this tool into our TSM codemods, which is currently difficult because of some invocation constraints:

1. Only one input path may be provided.
2. The input path must be a directory, not a file.
3. The input path must be relative to the current working directory, not an absolute path. (This one's just an accident; `path.join('/foo', '/bar')` yields `/foo/bar`, which isn't what you might expect.)

So, in this change, we upgrade `update_eslint_disable_lines.js` to accept multiple input paths, which may be files or directories, and may be relative or absolute.

Test Plan:
```
./update_eslint_disable_lines.js javascript/404-package javascript/about-package/about.js
```

Reviewers: kevinb, john, csilvers

Reviewed By: john, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D41195