Stop linting files on commit
Summary:
In accordance with our plans to integrate Prettier auto-formatting into our lint workflow, we're now linting on `git push` and `arc diff`, and no longer need to lint on commit. This change removes the actual file-linting logic from the `commit-msg` git hook, hooray!

This change will be deployed on Saturday, in a batch with other tooling changes and the JS codemod deploy to production, to ensure that folks developing in webapp have a consistent migration experience from a totally-non-Prettier world to a totally-Prettier world.

Test Plan:
In webapp, make a commit with a good message. Confirm that the commit is permitted.

In webapp, make a commit with a bad message (e.g. missing a Test Plan). Confirm that the commit is rejected.

In webapp, make a commit with a lint error in an arbitrary file. Confirm that the commit is permitted. (For good measure, try a `git push --dry-run`, to confirm that //pushing// the commit is rejected.)

Reviewers: csilvers, john

Reviewed By: csilvers, john

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D41357