Make the kotlin linter compatible with both java 8 and 9
Summary:

When we added a linter for kotlin, the latest version of java was (and
had been for a while) java 8. Since then, java 9 was released, and it
adds a bunch of extra warnings about using java reflection against
internal java APIs. On some people's systems, java 9 is now installed
either by default or for the purposes of other software. Unfortunately
the kotlin stanard library and the linter use these now-illegal
reflection operations, and our linter inteprets these as unparseable
lint errors.

It's not super easy to remove these warnings, as they may be
interleaved with other lint output, and I don't want to make any
assumptions about which of java 8 or 9 someone has installed (in
particular, android development might require a specific version, and
java is used for tons of other software someone might have installed).
Therefore, this commit adds a workaround for making the linter
compatible with both java 8 and java 9. In java 9, there are new
command-line flags that allow you to whitelist reflection against
certain internal packages, so I've added these flags to whitelist the
packages that the linter uses. Unfortunately, these flags error on java
8, so I've had to conditionally add them only on java 9.

Test plan:
- install java 8
- in `webapp/dataflow/log-export`: `ka-lint src`
- verify there are no errors
- uninstall java 8, install java 9
- in `webapp/dataflow/log-export`: `ka-lint src`
- verify there are still no errors

Auditors: ragini, csilvers