[lint on push] Copy changes
Summary:
I've been dogfooding the new lint-on-push behavior, and noticed some confusing output (or lack thereof).

In this change, I make some minor copy edits: When lint-on-push starts, we print a message including the number of files to lint (to inform why there's lag, and help the user spot-check whether it's a surprising number of files). Additionally, on commit, we now use a message that very clearly indicates that we //only// linted the commit message, not the change overall.

Test Plan: Make a commit, and see the new `commit-msg` copy. Make a push, and see the new lint-on-push copy.

Reviewers: csilvers, john

Reviewed By: csilvers

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D41416