Be more permissive with the regular expression for parsin the java version
Summary:

I added a regex to the kotlin linter for checking the dataflow version.
This correctly identifies the java version if running the oracle
version of java, but jenkins is running openjdk, so this fails. This
commit makes the regexp more permissive for identifying the java
version so that we match both the oracle and openjdk version strings.

Test Plan:
- `ka-lint ~/khan/webapp/dataflow/log-export/src`
- verify this completes successfully

Auditors: mattdunnrankin