[lint on push] Don't crash when there's nothing to push
Summary:
Whoops, in our previous implementation of lint-on-push, we crashed when the STDIN to our git hook was empty.

In this change, we solve the issue by skipping empty lines in STDIN.

Test Plan:
`git push` while the remote is already up-to-date, and see the usual git response (with no crashes).

`git push` while the remote is not yet up-to-date, and see linting and pushing.

Reviewers: csilvers, john

Reviewed By: csilvers

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D41415