[tsm codemods] Include blank lines as part of the eslint-disable "header"
Summary:
After running the `tsm-2017-js-codemods` tool, we noticed that some files stopped passing lint, because the file started with blank lines. This happened because the final upgrade, which calls `update_eslint_disable_lines.js`, removed some `eslint-disable` headers entirely - but the original file had a blank line after the header, which was preserved.

In this change, we update our line-filtering code to count blank lines as part of the header, if we haven't yet encountered any non-blank non-header lines.

Test Plan:
Edit `404-page.jsx` to add a redundant `var a;` to `javascript/404-package/404-page.jsx`.

Run `update_eslint_disable_lines.js javascript/404-package/404-page.jsx`, and see a top-of-file header get added.

Edit `404-page.jsx` to add a redundant empty line below the header, and remove the `var a;`.

Run `update_eslint_disable_lines.js javascript/404-package/404-page.jsx`, and see the top-of-file header //and// subsequent empty line be removed! (Before this change, the top-of-file header would be removed, but not the extra line.)

Reviewers: csilvers, john, kevinb

Reviewed By: csilvers

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D41417