Make the linter always emit utf-8, regardless of the console encoding.
This is necessary in python2, where if you try `print my_unicode_string`
it will give an error.  This became a problem when some of the linters
(*cough*prettier*cough*) started giving unicode output in its lint
messages.

In python3 life is more complicated, and I don't fully understand it.
In general python3 will let you print unicode, and it will encode it
the way the terminal wants to.  For me that's just ascii and things
break.  (I have no idea why my terminal doesn't advertise itself as
knowing utf-8.)  But the important issue is this code may be used in a
pipe, or a StringIO, in which case I have no idea how python3 picks
the encoding.  I figure it's better to just explicitly make it utf-8.
Nobody has a terminal that supports a different encoding anyway.

I am not happy with how hacky this is though. :-(

Test Plan:
In webapp, I changed my `.eslintrc` to read
        "./third_party/khan-linter-src/eslintrc.prettier",
instead of
        "./third_party/khan-linter-src/eslintrc",
and then I ran
   python2 ../devtools/khan-linter/runlint.py flow-typed
   python3 ../devtools/khan-linter/runlint.py flow-typed
In both cases it emitted some nice pretty utf-8 encoded text.

I then ran
   tools/runtests.py testing.lint_test.Lint_flow__typed.test
and it printed that same text again, this time having gone through a
pipe.