Get rid of the pyflakes exception about redefined functions.
Summary:
When we upgraded pyflakes, its error-reporting around redefined
functions improved, to the point that we can now trust it.

That is very nice, since sometimes we actually redefine a function (by
accident) and it would be nice if pyflakes could detect that.

Test Plan:
I ran
    ka-lint
in the webapp directory, and saw only legitimate errors (that I'll
clean up before landing).

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D41807