Upgrade ktlint to 0.15.0
Summary:

The latest version of the kotlin linter includes a bunch of extra
checks, as well as support for making IntelliJ show the same lint
errors and formatting conventions that it's going to enforce, which is
a huge time saver.

This commit upgrades to 0.15.0, which is the latest version.
There's an accompanying webapp commit that will have to be deployed
before landing this that includes all the lint fixes for this upgrade.

Test Plan:
- see test plan for accompanying webapp commit

Reviewers: benkraft, csilvers