Fix the json-check to better notice whether we're in a docstring.
Summary:
We have an exception to the 80-col line limit for json-like strings
within docstrings.  But that exception was firing for code that was
*not* in docstrings, due to being confused by one-line docstrings.

I fixed that up.  It was complex enough I refactored that logic into
its own helper function.

Test Plan:
I ran
   ka-lint
in webapp and found a few 80-col violations that had been skipped over
before.

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D41931