Accept `NoQA` as an alias for `Nolint`.
Summary:
This is part of our move to support flake8 more natively, which uses
`NoQA` syntax.  I wish it supported `@NoQA`, but we'd have to fork it
for that, and it's probably not worth it.  Still, sad.

Test Plan:
I ran
   make lint
on a webapp with a bunch of Nolint's replaced by NoQA's, and it passed lint.

Reviewers: colin

Subscribers: kevinb, benkraft

Differential Revision: https://phabricator.khanacademy.org/D42637