Have UnicodeDecodeError be one of the reasons that a file-read can fail.
This can't happen in python2 but it can in python3, if you try to lint
a binary file (which will happen if you have an image file in your
repo, say, since we still lint it with GitLint).  Before we'd die, now
we just print a cryptic error.

Auditors: colin

Test Plan:
Ran
   LC_ALL=C ka-lint
in the khanalytics repo, and saw it skip
core/monitor/src/back-button.jsx rather than crashing entirely.