Add yaml parsing to khan-linter.
Webapp had a linter that checked that a yaml file is syntactically
correct (no syntax errors trying to load it).  That doesn't seem
webapp-specific to me!  Let's do that for yaml files in *every* repo!

My biggest worry here is that we're checking in `.so` files to
vendor/, which probably won't work for most users unless they have the
same architecture I do.  But I checked the yaml source and it seems
like it's good about falling back to pure python if it can't load the
`.so` file for some reason.

Test Plan:
I changed webapp's queue.yaml to unindent a line, and ran
   ka-lint queue.yaml
and it said:
   queue.yaml:35:1: E=yaml= Error parsing yaml: expected <block end>, but found '<block mapping start>' while parsing a block mapping

The webapp linter (which is still there) said:
   queue.yaml:35: E314 Error parsing yaml: expected <block end>, but found '<block mapping start>' while parsing a block mapping. (yaml_lint.lint_yamls_are_parseable)

so it looks like we're matching pretty well.  (I'm not sure the column
number is needed, but it doesn't hurt.)