Manually fix a bug in the `backports` vendored package.
According to https://pypi.python.org/pypi/backports
`backports/__init__.py` needs to be set up in a certain way.
But when we do
    pip install -t vendor/py2 configparser
it does not set up that `__init__.py` file!  So we do it manually.

Resolves INFRA-860

Auditors: sean

Test Plan:
I put a breakpoint in linters.py right after the `import yaml` line,
and ran
   ka-lint linters.py
at the pdb prompt I did
   (Pdb) import backports.configparser
   (Pdb) backports.configparser.__file__
and it said vendor/py2/something!  I also did
   (Pdb) import sys
   (Pdb) sys.modules['backports'].__path__
and it had two entries, one for my virtualenv and one for vendor/py2.