Force flake8 to run in serial so it doesn't leak processes.
A bug in flake8 causes it to leak processes the way we use it in
linters.py.  This is normally fine -- the OS will harvest the
processes when ka-lint finishes -- but is bad inside webapp tests
because those leaked processes stick around throughout all of
runtests.py, and worse they do a lot of sleeping, which advances the
fake-datetime a ton.  This causes all manner of tests to fail in
surprising ways.

Running with jobs=1 matches our old behavior anyway.  It doesn't
matter on jenkins, where our machines have only 1 CPU and we were
always running with jobs=1 anyway.  And it doesn't matter for webapp
tests, since we parallelize at the runtests.py level.  And I don't
*think* it matters for `arc lint`, which runs ka-lint one file at a
time in any case.  So it's probably a noop.

Auditors: colin, benkraft

Test plan:
Will update webapp substate and run `make check`