Use six.moves.range instead of xrange
Summary:

26f6d6f8c1ef4470701c0087709e80b0994a95df added iterative linting over
javascript files to avoid an issue where we tried to pass too many
arguments to the linter. This introduced a usage of `xrange` which is
python2-only. This commit changes it so we use `six.moves.range` which
acts as `xrange` in python2 and `range` in python3.

I also fixed a lint error in linters.py (using `stdout` which is not
defined, instead of `stdout_lines` I presume) that showed up during the
test plan.

Test Plan:
- `make lint` in the khanalytics repo with a python 3 virtualenv, see
  that the linter passes
- `ka-lint` in the khan-linter repo with a python 2 virtualenv, see
  that the linter passes

Auditors: csilvers