Report on eslint errors even if they refer to an invalid line
This can happen e.g. if the file in question is empty, but it reports an
error on line 1.  We do the best we can; namely we pretend the line was
empty and hope that's good enough to be able to report the error to the
user.

Test plan: crossed fingers

Auditors: john, csilvers