[flow upgrade] Update khan-linter's eslint-plugin-react to 7.7.0
Summary:
In preparation for upgrading Flow to 0.53.0, we need to update our React ESLint plugin. This is because Flow 0.53.0 requires a new syntax for declaring Flow types of React components, which our current version of the plugin doesn't know about, but the latest version does.

Without this change, we see many spurious `Ereact/prop-types` errors in webapp, asserting that `this.props.title` doesn't refer to a real prop - but only because it doesn't know the new Flow prop type syntax. After this change, those errors are no longer reported.

This change //does// yield //new// errors in webapp: the plugin is now detecting genuine errors in legacy components that it previously wasn't able to detect. In D43333, we preemptively add `eslint-disable` suppressions for those errors, in anticipation of this `khan-linter` upgrade.

NOTE: This change MUST NOT land until D43333 is deployed to `master` on `webapp`.

NOTE: This commit does not include the updated contents of `node_modules`, to ease reviewing. Before landing this change, run `npm install`, and commit the resulting changes.

Test Plan:
First, check out `mattdr-flow-testing` in webapp, in which I've installed Flow 0.53.0, and codemodded our React components to use the new Flow syntax.

Then, run the old version of the linter in webapp. See lots of spurious `Ereact/prop-types` errors, oh no!

Then, run `npm install` in `khan-linter` to install the new version `eslint-plugin-react`.

Finally, run the linter in webapp again. The spurious `Ereact/prop-types` errors should be gone (though there are still many unresolved lint errors on that branch).

Reviewers: csilvers, john, miguel

Reviewed By: csilvers, john

Subscribers: kevinb, colin, amy

Differential Revision: https://phabricator.khanacademy.org/D43336